backend/requirements.txt
Flask==3.1.0
Werkzeug==3.1.3
Flask-SQLAlchemy
Flask-JWT-Extended==4.7.1
Flask-Bcrypt
PyMySQL
python-dotenv
requests
stripe
Flask-Cors
gunicorn==23.0.0
markdown
flask==2.2.3
flask-cors==3.0.10
flask-sqlalchemy==3.0.3
flask-jwt-extended==4.4.4
flask-bcrypt==1.0.1
pymysql==1.0.3
python-dotenv==1.0.0
pillow==9.5.0
werkzeug==2.2.3
stripe==5.4.0